12_OpenAI_Agent_SDK/10_Context_Mangement/pyproject.toml
[project]
name = "10-context-mangement"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai-agents>=0.1.0",
    "python-dotenv>=1.1.1",
]
